+++ src/main/java/uk/ac/ed/ilp/model/requests/NextPositionRequest.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class NextPositionRequest {
     private LngLat start;
-    private int angle; // degrees
+    private Double angle; // degrees
 }
